requirements.txt
defusedxml==0.4.1
Django==1.7.11
django-debug-toolbar==1.4
django-djconfig==0.1.8
django-filter==0.9.2
django-guardian==1.4.2
django-haystack==2.3.2
django-infinite-scroll-pagination==0.1.3
django-rosetta==0.7.8
django-social-auth==0.7.28
django-userena==1.0.1
django-versatileimagefield==1.3
djangorestframework==3.1.1
easy-thumbnails==2.3
httplib2==0.9.2
Markdown==2.6.2
microsofttranslator==0.7
mistune==0.4.1
nltk==3.1
oauth2==1.9.0.post1
oauthlib==1.0.3
Pillow==3.1.1
polib==1.0.7
psycopg2==2.6.1
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
South==1.0.2
sqlparse==0.1.19
Whoosh==2.5.7



#Required to Pillow
#sudo apt-get install libjpeg-dev


#sudo apt-get install postfix
#sudo apt-get install aptitude
#sudo aptitude install apache2 apache2.2-common apache2-mpm-prefork apache2-utils libexpat1 ssl-cert

#install wsgi
#sudo aptitude install libapache2-mod-wsgi
#sudo easy_install django-rosetta